+++ CHANGES.md
@@ -1,3 +1,7 @@
+
+
+* Add `Fmt.if'`, thanks to Rizo for the suggestion and patch.
+
 v0.9.0 2021-10-22 Zagreb
 ------------------------