+++ CHANGES.md
@@ -1,5 +1,5 @@
 
-
+* Install each library in its own directory.
 * Add `Fmt.if'`, thanks to Rizo for the suggestion and patch.
 
 v0.9.0 2021-10-22 Zagreb